requirements.txt
requests >= 2.32.3, < 3.0.0
python-dotenv >= 1.0.1, < 2.0.0
